Day 12: Part 2 complete
Switch to a recursive solution for simplicity and efficiency. Even
bulk-skipping '.' characters, it was still far too slow. Memoize
the results and it's pretty darn fast (if awkward in Kotlin).

Fixed the incorrect parsing for part 2 (missing added '?').

Switch group lengths to Int to avoid casting at all indexing points.